+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/PhoenixApplication.kt
@@ -21,9 +21,11 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
+import fr.acinq.phoenix.PhoenixGlobal
 import fr.acinq.phoenix.android.utils.Logging
 import fr.acinq.phoenix.android.utils.SystemNotificationHelper
 import fr.acinq.phoenix.android.utils.datastore.GlobalPrefs
+import fr.acinq.phoenix.utils.PlatformContext
 import org.slf4j.LoggerFactory
 
 
@@ -35,10 +37,12 @@ class PhoenixApplication : Application() {
     private val log = LoggerFactory.getLogger(this::class.java)
 
     lateinit var globalPrefs: GlobalPrefs
+    lateinit var phoenixGlobal: PhoenixGlobal
 
     override fun onCreate() {
         super.onCreate()
 
+        phoenixGlobal = PhoenixGlobal(PlatformContext(applicationContext))
         globalPrefs = GlobalPrefs(applicationContext.globalPrefs)
         BusinessManager.initialize(applicationContext)
