+++ phoenix-ios/phoenix-notifySrvExt/PhoenixManager.swift
@@ -32,7 +32,7 @@ class PhoenixManager {
 	
 	private var business: PhoenixBusiness? = nil
 	private var oldBusiness: PhoenixBusiness? = nil
-	
+
 	private var walletId: WalletIdentifier? = nil
 
 	private var cancellables = Set<AnyCancellable>()
@@ -104,14 +104,14 @@ class PhoenixManager {
 	}
 	
 	public func groupPrefs() -> GroupPrefs_Wallet? {
-		
+
 		if let walletId {
 			return GroupPrefs.wallet(walletId)
 		} else {
 			return nil
 		}
 	}
-	
+
 	public func exchangeRate(fiatCurrency: FiatCurrency) -> ExchangeRate.BitcoinPriceRate? {
 		
 		return Utils.exchangeRate(for: fiatCurrency, fromRates: fiatExchangeRates)
@@ -144,12 +144,12 @@ class PhoenixManager {
 				
 				var sealedBox: SealedBox_ChaChaPoly? = nil
 				var id: String? = nil
-				
+
 				switch securityFile {
 				case .v0(let v0):
 					sealedBox = v0.keychain
 					id = KEYCHAIN_DEFAULT_ID
-					
+
 				case .v1(let v1):
 					if let target {
 						sealedBox = v1.wallets[target]?.keychain
@@ -159,12 +159,12 @@ class PhoenixManager {
 						id = defaultTarget
 					}
 				}
-				
+
 				guard let sealedBox, let id else {
 					unlockWithRecoveryPhrase(nil)
 					return
 				}
-				
+
 				let keychainResult = SharedSecurity.shared.readKeychainEntry(id, sealedBox)
 				switch keychainResult {
 				case .failure(_):
@@ -209,32 +209,32 @@ class PhoenixManager {
 			passphrase: ""
 		)
 		let walletInfo = business.walletManager.loadWallet(seed: seed)
-		
+
 		let wid = WalletIdentifier(chain: business.chain, walletInfo: walletInfo)
 		walletId = wid
-		
+
 		if let targetNodeIdHash {
 			guard targetNodeIdHash == wid.nodeIdHash else {
 				log.warning("unlock(): ignoring: target.nodeIdHash != unlocked.nodeIdHash")
 				return
 			}
 		}
-		
+
 		let groupPrefs = GroupPrefs.wallet(wid)
-		
+
 		if let electrumConfigPrefs = groupPrefs.electrumConfig {
 			business.appConfigurationManager.updateElectrumConfig(config: electrumConfigPrefs.customConfig)
 		} else {
 			business.appConfigurationManager.updateElectrumConfig(config: nil)
 		}
-		
+
 		let primaryFiatCurrency = groupPrefs.fiatCurrency
-		let preferredFiatCurrencies = AppConfigurationManager.PreferredFiatCurrencies(
+		let preferredFiatCurrencies = PreferredFiatCurrencies(
 			primary: primaryFiatCurrency,
 			others: groupPrefs.preferredFiatCurrencies
 		)
-		business.appConfigurationManager.updatePreferredFiatCurrencies(
-			current: preferredFiatCurrencies
+		business.currencyManager.monitorCurrencies(
+			value: preferredFiatCurrencies
 		)
 
 		let startupParams = StartupParams(
