+++ phoenix-shared/src/androidMain/kotlin/fr/acinq/phoenix/PhoenixGlobalBusinessAndroid.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2025 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix
+
+import android.content.Context
+import fr.acinq.lightning.logging.LoggerFactory
+import fr.acinq.phoenix.utils.PlatformContext
+import fr.acinq.phoenix.utils.logger.PhoenixLoggerConfig
+
+//actual class PhoenixGlobalBusiness(context: Context) {
+//    private val applicationContext = context.applicationContext
+//
+//    actual val loggerFactory: LoggerFactory = LoggerFactory(PhoenixLoggerConfig(PlatformContext(applicationContext)))
+//
+//
+//
+//
+//
+//}
