+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/PhoenixGlobalBusiness.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2025 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix
+
+import fr.acinq.lightning.logging.LoggerFactory
+import fr.acinq.lightning.logging.info
+import fr.acinq.phoenix.db.SqliteAppDb
+import fr.acinq.phoenix.db.createAppDbDriver
+import fr.acinq.phoenix.managers.fiatcurrencies.CurrencyManager
+import fr.acinq.phoenix.managers.NetworkMonitor
+import fr.acinq.phoenix.utils.PlatformContext
+import fr.acinq.phoenix.utils.logger.PhoenixLoggerConfig
+
+
+class PhoenixGlobal(val ctx: PlatformContext) {
+
+    val loggerFactory = LoggerFactory(PhoenixLoggerConfig(ctx))
+    private val logger = loggerFactory.newLogger(this::class)
+
+    val appDb by lazy { SqliteAppDb(createAppDbDriver(ctx)) }
+
+    val networkMonitor by lazy { NetworkMonitor(loggerFactory, ctx) }
+
+    val currencyManager by lazy { CurrencyManager(loggerFactory, appDb) }
+
+    init {
+        logger.info { "init PhoenixGlobal..." }
+    }
+}
