+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/managers/fiatcurrencies/apis/YadioApi.kt
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2025 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix.managers.fiatcurrencies.apis
+
+import fr.acinq.lightning.logging.LoggerFactory
+import fr.acinq.lightning.logging.error
+import fr.acinq.phoenix.data.ExchangeRate
+import fr.acinq.phoenix.data.FiatCurrency
+import fr.acinq.phoenix.data.YadioResponse
+import io.ktor.client.request.get
+import io.ktor.client.statement.HttpResponse
+import io.ktor.client.statement.bodyAsText
+import kotlinx.datetime.Clock
+import kotlin.time.Duration.Companion.minutes
+
+/**
+ * The yadio API is used to fetch various "free market" prices.
+ * For example:
+ * - CUP => government controlled exchange rate
+ * - CUP_FM => free market exchange rate
+ */
+class YadioAPI(loggerFactory: LoggerFactory) : ExchangeRateApi {
+
+    val log = loggerFactory.newLogger(this::class)
+
+    override val name = "yadio"
+    override val refreshDelay = 120.minutes
+    override val fiatCurrencies = setOf(FiatCurrency.CUP_FM, FiatCurrency.LBP_BM)
+
+    override suspend fun fetch(targets: Set<FiatCurrency>): List<ExchangeRate> {
+        val httpResponse: HttpResponse? = try {
+            ExchangeRateApi.httpClient.get(urlString = "https://api.yadio.io/exrates/USD")
+        } catch (e: Exception) {
+            log.error { "failed to get exchange rates from api.yadio.io: $e" }
+            null
+        }
+        val parsedResponse: YadioResponse? = httpResponse?.let {
+            try {
+                ExchangeRateApi.json.decodeFromString<YadioResponse>(httpResponse.bodyAsText())
+            } catch (e: Exception) {
+                log.error { "failed to get exchange rates response from api.yadio.io: $e" }
+                null
+            }
+        }
+
+        val timestampMillis = Clock.System.now().toEpochMilliseconds()
+        val fetchedRates: List<ExchangeRate> = parsedResponse?.let {
+            targets.mapNotNull { fiatCurrency ->
+                val name = fiatCurrency.name.take(3)
+                parsedResponse.usdRates[name]?.let { valueAsDouble ->
+                    ExchangeRate.UsdPriceRate(
+                        fiatCurrency = fiatCurrency,
+                        price = valueAsDouble,
+                        source = "yadio.io",
+                        timestampMillis = timestampMillis
+                    )
+                }
+            }
+        } ?: listOf()
+
+        return fetchedRates
+    }
+}
