+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/utils/MetadataQueue.kt
@@ -2,12 +2,9 @@ package fr.acinq.phoenix.utils
 
 import fr.acinq.lightning.utils.UUID
 import fr.acinq.phoenix.data.WalletPaymentMetadata
-import fr.acinq.phoenix.managers.CurrencyManager
 import kotlinx.coroutines.flow.MutableStateFlow
 
-class MetadataQueue(
-    private val currencyManager: CurrencyManager?
-) {
+class MetadataQueue() {
 
     private var metadataQueue = MutableStateFlow(mapOf<UUID, WalletPaymentMetadata>())
 
@@ -32,12 +29,6 @@ class MetadataQueue(
         metadataQueue.value = newMap
 
         val row = oldMap[id] ?: WalletPaymentMetadata()
-
-        // Append the current exchange rate, unless it was explicitly set earlier.
-        return if (row.originalFiat != null) {
-            row
-        } else {
-            row.copy(originalFiat = currencyManager?.calculateOriginalFiat())
-        }
+        return row
     }
 }
